backend/requirements.txt
# Core Django
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# HTTP Requests
requests==2.31.0

# WebSockets (optional for now)
channels==4.0.0
daphne==4.0.0
channels-redis==4.2.0

# F1 Data Libraries
fastf1==3.3.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Environment variables
python-dotenv==1.0.0
